move fallenBehindConfig
