+++ platform-sdk/consensus-gossip-impl/src/test/java/org/hiero/consensus/gossip/impl/gossip/shadowgraph/RpcShadowgraphSynchronizerTest.java
@@ -68,8 +68,11 @@ class RpcPeerHandlerTest {
 
         this.syncMetrics = mock(SyncMetrics.class);
         this.selfId = NodeId.of(1);
+        // The value used in this test does not matter. The fallen behind manager is only used to determine if
+        // SELF or OTHER is behind based on exchanged event windows.
+        final double fallenBehindThreshold = 1.0;
         this.fallenBehindManager = new FallenBehindMonitor(
-                RosterFactory.randomRoster(new Random(), NUM_NODES), configuration, new NoOpMetrics(), selfId);
+                RosterFactory.randomRoster(new Random(), NUM_NODES), new NoOpMetrics(), selfId, fallenBehindThreshold);
         this.eventHandler = mock(Consumer.class);
         this.gossipSender = mock(GossipRpcSender.class);
         this.syncProgressReporter = mock(Consumer.class);
