+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/config/PlatformConfigurationExtension.java
@@ -13,6 +13,7 @@ import com.swirlds.platform.monitor.StatusMonitorWiringConfig;
 import com.swirlds.platform.uptime.UptimeConfig;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Set;
+import org.hiero.consensus.FallenBehindConfig;
 import org.hiero.consensus.model.node.NodeId;
 
 /**
@@ -31,6 +32,7 @@ public class PlatformConfigurationExtension implements ConfigurationExtension {
         // Load Configuration Definitions
         return Set.of(
                 ModulesConfig.class,
+                FallenBehindConfig.class,
                 OSHealthCheckConfig.class,
                 PlatformMetricsConfig.class,
                 StatusMonitorWiringConfig.class,
