+++ platform-sdk/swirlds-platform-core/src/main/java/org/hiero/consensus/FallenBehindConfig.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.consensus.config;
+package org.hiero.consensus;
 
 import com.swirlds.config.api.ConfigData;
 import com.swirlds.config.api.ConfigProperty;
